Consolidate parsing of string literals, and allow \b and \f
Regular expressions governing the validity of strings, raw strings
and byte strings are now consolidated in `edgeql.parser.grammar.lexutils`.

Additionally, the `\b` and `\f` escape sequences are now recognized as
they are in PostgreSQL and most common languages.
